feat!: rename Passage struct and ctor signature and remove old Error class
